refactor

+++ includes/winutils.hpp
@@ -7,7 +7,7 @@ int DisplayErrorMessageBox(const char* error);
 std::string GetExecutableDirectory();
 std::vector<std::string> getDLLs();
 Environment IncludeLIB(std::string lib, Environment &env);
-void sleep(int idk);
+void sleep(double idk);
 
 #include <ECMOD.hpp>
